# services/trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing portfolios, executing trades, and handling financial market operations.",
        "title": "Trading Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/actuary/{actId}/assets/funds": {
            "get": {
                "description": "Returns all investment funds managed by the specified actuary (supervisor). Shows fund value and liquidity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Get funds managed by an actuary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Actuary ID",
                        "name": "actId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ActuaryFundResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/client/{clientId}/funds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all investment fund positions for the specified client, including share percentage and current value in RSD.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Get client's investment fund positions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FundPositionSummaryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/funds": {
            "get": {
                "description": "Returns a paginated list of all investment funds with optional filtering and sorting. Accessible to actuaries and clients.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Get all investment funds",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by fund name (case-insensitive substring)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field: name, minimum_contribution, created_at, liquid_assets",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc, desc",
                        "name": "sort_dir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListFundsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/investment-funds": {
            "post": {
                "description": "Supervisor creates a new investment fund. An RSD account is automatically created for the fund.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Create a new investment fund",
                "parameters": [
                    {
                        "description": "Fund details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFundRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFundResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/investment-funds/{fundId}": {
            "get": {
                "description": "Retrieves detailed information about an investment fund, including holdings, current value, performance history, and account balance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Get investment fund details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fund ID",
                        "name": "fundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FundDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an investment fund by ID. Only supervisors can delete funds. Fails if the fund has active client positions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Delete an investment fund",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fund ID",
                        "name": "fundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/investment-funds/{fundId}/invest": {
            "post": {
                "description": "Allows a client or supervisor to invest money into an investment fund.\nClients must provide one of their own accounts; supervisors must provide a bank account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Invest into a fund",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fund ID",
                        "name": "fundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Investment details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InvestInFundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InvestInFundResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/investment-funds/{fundId}/withdraw": {
            "post": {
                "description": "Allows a client or supervisor to withdraw money from an investment fund position.\nClients must provide one of their own accounts; supervisors must provide a bank account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment-funds"
                ],
                "summary": "Withdraw from a fund",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fund ID",
                        "name": "fundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Withdrawal details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawFromFundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawFromFundResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/profit/funds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all investment funds with bank share, manager info and profit calculation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profit"
                ],
                "summary": "Get investment fund positions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FundPositionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ActuaryFundResponse": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fund_id": {
                    "type": "integer"
                },
                "fund_value": {
                    "type": "number"
                },
                "liquid_assets": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateFundRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "dividend_reinvestment_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "minimum_contribution": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateFundResponse": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dividend_reinvestment_percent": {
                    "type": "number"
                },
                "fund_id": {
                    "type": "integer"
                },
                "manager_id": {
                    "type": "integer"
                },
                "minimum_contribution": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.FundDetailResponse": {
            "type": "object",
            "properties": {
                "account_balance": {
                    "type": "number"
                },
                "annual_return": {
                    "type": "number"
                },
                "average_market_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FundPerformanceEntry"
                    }
                },
                "description": {
                    "type": "string"
                },
                "dividend_reinvestment_percent": {
                    "type": "number"
                },
                "fund_value": {
                    "type": "number"
                },
                "holdings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SecurityHoldingResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "manager": {
                    "type": "string"
                },
                "max_drawdown": {
                    "type": "number"
                },
                "min_investment": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "performance_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FundPerformanceEntry"
                    }
                },
                "profit": {
                    "type": "number"
                },
                "reward_to_variability": {
                    "type": "number"
                },
                "volatility": {
                    "type": "number"
                }
            }
        },
...
        "dto.FundPositionSummaryResponse": {
            "type": "object",
            "properties": {
                "clients_share_percent": {
                    "type": "number"
                },
                "clients_share_value_rsd": {
                    "type": "number"
                },
                "fund_description": {
                    "type": "string"
                },
                "fund_id": {
                    "type": "integer"
                },
                "fund_name": {
                    "type": "string"
                },
                "total_profit": {
                    "type": "number"
                }
            }
        },
...
        "dto.InvestInFundRequest": {
            "type": "object",
            "required": [
                "account_number",
                "amount"
            ],
            "properties": {
                "account_number": {
                    "description": "AccountNumber is the source account for the investment.\nFor clients this must be one of their own accounts.\nFor supervisors this must be a bank account.",
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.ListFundsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FundSummaryResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}